backend/requirements.txt
Django==1.11.4
django-disposable-email-checker==1.2.1
python-dateutil==2.6.1
pymongo==3.6.0
mock==2.0.0
factory-boy==2.9.2
prospector==0.12.7
bandit==1.4.0
pytest==3.2.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-pythonpath==0.7.1
pytest-xdist==1.20.0
ptpython==0.41
amqp==2.2.2
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
django-braces==1.11.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.3
flower==0.9.2
gunicorn==19.7.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
Markdown==2.6.9
msgpack-python==0.4.8
oauthlib==2.0.6
psycopg2==2.7.3.2
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
zope.interface==4.4.3
